requirements.txt
pytorch-lightning==2.0.0
omegaconf==2.3.0
jaxtyping
typeguard
diffusers<0.20
transformers==4.28.1
accelerate
opencv-python
tensorboard
matplotlib
imageio>=2.28.0
imageio[ffmpeg]
libigl
xatlas
trimesh[easy]
networkx
pysdf
PyMCubes
wandb
gradio==4.11.0
torchmetrics
einops
git+https://github.com/ashawkey/envlight.git
git+https://github.com/NVlabs/nvdiffrast.git
